+++ rasa/nlu/tokenizers/whitespace_tokenizer.py
@@ -3,8 +3,10 @@ from typing import Any, Dict, List, Text
 import regex
 import re
 
+from rasa.constants import DOCS_URL_COMPONENTS
 from rasa.nlu.tokenizers.tokenizer import Token, Tokenizer
 from rasa.nlu.training_data import Message
+import rasa.utils.common as common_utils
 
 
 class WhitespaceTokenizer(Tokenizer):
@@ -16,22 +18,28 @@ class WhitespaceTokenizer(Tokenizer):
         "intent_split_symbol": "_",
         # Regular expression to detect tokens
         "token_pattern": None,
-        # Text will be tokenized with case sensitive as default
-        "case_sensitive": True,
     }
+
+    # the following language should not be tokenized using the WhitespaceTokenizer
+    not_supported_language_list = ["zh", "ja", "th"]
 
     def __init__(self, component_config: Dict[Text, Any] = None) -> None:
         """Construct a new tokenizer using the WhitespaceTokenizer framework."""
 
         super().__init__(component_config)
 
-        self.case_sensitive = self.component_config["case_sensitive"]
-
         self.emoji_pattern = self.get_emoji_regex()
+
+        if "case_sensitive" in self.component_config:
+            common_utils.raise_warning(
+                "The option 'case_sensitive' was moved from the tokenizers to the "
+                "featurizers.",
+                docs=DOCS_URL_COMPONENTS,
+            )
 
     @staticmethod
     def get_emoji_regex():
-        emoji_pattern = re.compile(
+        return re.compile(
             "["
             "\U0001F600-\U0001F64F"  # emoticons
             "\U0001F300-\U0001F5FF"  # symbols & pictographs
@@ -39,20 +47,23 @@ class WhitespaceTokenizer(Tokenizer):
             "\U0001F1E0-\U0001F1FF"  # flags (iOS)
             "\U00002702-\U000027B0"
             "\U000024C2-\U0001F251"
+            "\u200d"  # zero width joiner
+            "\u200c"  # zero width non-joiner
             "]+",
             flags=re.UNICODE,
         )
-        return emoji_pattern
 
     def remove_emoji(self, text: Text) -> Text:
+        """Remove emoji if the full text, aka token, matches the emoji regex."""
+        match = self.emoji_pattern.fullmatch(text)
 
-        return self.emoji_pattern.sub(r"", text)
+        if match is not None:
+            return ""
+
+        return text
 
     def tokenize(self, message: Message, attribute: Text) -> List[Token]:
         text = message.get(attribute)
-
-        if not self.case_sensitive:
-            text = text.lower()
 
         # we need to use regex instead of re, because of
         # https://stackoverflow.com/questions/12746458/python-unicode-regular-expression-matching-failing-with-some-unicode-characters