+++ rasa/nlu/training_data/formats/markdown.py
@@ -47,7 +47,7 @@ ESCAPE = re.compile(r"[\b\f\n\r\t]")
 
 
 def encode_string(s: Text) -> Text:
-    """Return a encoded python string."""
+    """Return an encoded python string."""
 
     def replace(match: Match) -> Text:
         return ESCAPE_DCT[match.group(GROUP_COMPLETE_MATCH)]