+++ tests/core/test_domain.py
@@ -5,17 +5,14 @@ from pathlib import Path
 import pytest
 from _pytest.tmpdir import TempdirFactory
 
+from rasa.constants import DEFAULT_SESSION_EXPIRATION_TIME_IN_MINUTES
 from rasa.core.constants import (
     DEFAULT_KNOWLEDGE_BASE_ACTION,
     SLOT_LISTED_ITEMS,
     SLOT_LAST_OBJECT,
     SLOT_LAST_OBJECT_TYPE,
 )
-from rasa.core.domain import (
-    USED_ENTITIES_KEY,
-    USE_ENTITIES_KEY,
-    IGNORE_ENTITIES_KEY,
-)
+from rasa.core.domain import USED_ENTITIES_KEY, USE_ENTITIES_KEY, IGNORE_ENTITIES_KEY
 from rasa.core import training, utils
 from rasa.core.domain import Domain, InvalidDomain, SessionConfig
 from rasa.core.featurizers import MaxHistoryTrackerFeaturizer
@@ -167,7 +164,7 @@ def test_domain_from_template():
 
     assert not domain.is_empty()
     assert len(domain.intents) == 10
-    assert len(domain.action_names) == 13
+    assert len(domain.action_names) == 15
 
 
 def test_avoid_action_repetition():
@@ -237,7 +234,7 @@ def test_domain_fails_on_unknown_custom_slot_type(tmpdir, domain_unkown_slot_typ
 
 
 def test_domain_to_yaml():
-    test_yaml = """config:
+    test_yaml = f"""config:
   store_entities_as_slots: true
 entities: []
 forms: []
@@ -247,8 +244,8 @@ responses:
   - text: hey there!
 session_config:
   carry_over_slots_to_new_session: true
-  session_expiration_time: 60
-slots: {}"""
+  session_expiration_time: {DEFAULT_SESSION_EXPIRATION_TIME_IN_MINUTES}
+slots: {{}}"""
 
     domain = Domain.from_yaml(test_yaml)
     # python 3 and 2 are different here, python 3 will have a leading set
@@ -258,7 +255,7 @@ slots: {}"""
 
 
 def test_domain_to_yaml_deprecated_templates():
-    test_yaml = """actions:
+    test_yaml = f"""actions:
 - utter_greet
 config:
   store_entities_as_slots: true
@@ -270,10 +267,10 @@ templates:
   - text: hey there!
 session_config:
   carry_over_slots_to_new_session: true
-  session_expiration_time: 60
-slots: {}"""
+  session_expiration_time: {DEFAULT_SESSION_EXPIRATION_TIME_IN_MINUTES}
+slots: {{}}"""
 
-    target_yaml = """actions:
+    target_yaml = f"""actions:
 - utter_greet
 config:
   store_entities_as_slots: true
@@ -285,8 +282,8 @@ responses:
   - text: hey there!
 session_config:
   carry_over_slots_to_new_session: true
-  session_expiration_time: 60
-slots: {}"""
+  session_expiration_time: {DEFAULT_SESSION_EXPIRATION_TIME_IN_MINUTES}
+slots: {{}}"""
 
     domain = Domain.from_yaml(test_yaml)
     # python 3 and 2 are different here, python 3 will have a leading set
@@ -374,6 +371,30 @@ session_config:
     assert merged.session_config == SessionConfig(40, True)
 
 
+def test_merge_domain_with_forms():
+    test_yaml_1 = """
+    forms:
+    # Old style form definitions (before RulePolicy)
+    - my_form
+    - my_form2
+    """
+
+    test_yaml_2 = """
+    forms:
+    - my_form3:
+        slot1:
+          type: from_text
+    """
+
+    domain_1 = Domain.from_yaml(test_yaml_1)
+    domain_2 = Domain.from_yaml(test_yaml_2)
+    domain = domain_1.merge(domain_2)
+
+    expected_number_of_forms = 3
+    assert len(domain.form_names) == expected_number_of_forms
+    assert len(domain.forms) == expected_number_of_forms
+
+
 @pytest.mark.parametrize(
     "intents, entities, intent_properties",
     [
@@ -406,7 +427,7 @@ session_config:
             ],
             ["entity", "other", "third"],
             {
-                "greet": {"triggers": "utter_goodbye", USED_ENTITIES_KEY: ["entity"],},
+                "greet": {"triggers": "utter_goodbye", USED_ENTITIES_KEY: ["entity"]},
                 "goodbye": {USED_ENTITIES_KEY: ["entity", "other", "third"]},
             },
         ),
@@ -417,7 +438,7 @@ session_config:
             ],
             ["entity", "other", "third"],
             {
-                "greet": {USED_ENTITIES_KEY: [], "triggers": "utter_goodbye",},
+                "greet": {USED_ENTITIES_KEY: [], "triggers": "utter_goodbye"},
                 "goodbye": {USED_ENTITIES_KEY: []},
             },
         ),
@@ -430,29 +451,24 @@ def test_collect_intent_properties(intents, entities, intent_properties):
 def test_load_domain_from_directory_tree(tmpdir_factory: TempdirFactory):
     root = tmpdir_factory.mktemp("Parent Bot")
     root_domain = {"actions": ["utter_root", "utter_root2"]}
-    utils.dump_obj_as_yaml_to_file(root / "domain.yml", root_domain)
+    utils.dump_obj_as_yaml_to_file(root / "domain_pt1.yml", root_domain)
 
     subdirectory_1 = root / "Skill 1"
     subdirectory_1.mkdir()
     skill_1_domain = {"actions": ["utter_skill_1"]}
-    utils.dump_obj_as_yaml_to_file(subdirectory_1 / "domain.yml", skill_1_domain)
+    utils.dump_obj_as_yaml_to_file(subdirectory_1 / "domain_pt2.yml", skill_1_domain)
 
     subdirectory_2 = root / "Skill 2"
     subdirectory_2.mkdir()
     skill_2_domain = {"actions": ["utter_skill_2"]}
-    utils.dump_obj_as_yaml_to_file(subdirectory_2 / "domain.yml", skill_2_domain)
+    utils.dump_obj_as_yaml_to_file(subdirectory_2 / "domain_pt3.yml", skill_2_domain)
 
     subsubdirectory = subdirectory_2 / "Skill 2-1"
     subsubdirectory.mkdir()
     skill_2_1_domain = {"actions": ["utter_subskill", "utter_root"]}
     # Check if loading from `.yaml` also works
-    utils.dump_obj_as_yaml_to_file(subsubdirectory / "domain.yaml", skill_2_1_domain)
-
-    subsubdirectory_2 = subdirectory_2 / "Skill 2-2"
-    subsubdirectory_2.mkdir()
-    excluded_domain = {"actions": ["should not be loaded"]}
     utils.dump_obj_as_yaml_to_file(
-        subsubdirectory_2 / "other_name.yaml", excluded_domain
+        subsubdirectory / "domain_pt4.yaml", skill_2_1_domain
     )
 
     actual = Domain.load(str(root))
@@ -533,7 +549,7 @@ def test_check_domain_sanity_on_invalid_domain():
             slots=[],
             templates={},
             action_names=["random_name", "random_name"],
-            form_names=[],
+            forms=[],
         )
 
     with pytest.raises(InvalidDomain):
@@ -543,7 +559,7 @@ def test_check_domain_sanity_on_invalid_domain():
             slots=[TextSlot("random_name"), TextSlot("random_name")],
             templates={},
             action_names=[],
-            form_names=[],
+            forms=[],
         )
 
     with pytest.raises(InvalidDomain):
@@ -553,7 +569,7 @@ def test_check_domain_sanity_on_invalid_domain():
             slots=[],
             templates={},
             action_names=[],
-            form_names=[],
+            forms=[],
         )
 
     with pytest.raises(InvalidDomain):
@@ -563,7 +579,7 @@ def test_check_domain_sanity_on_invalid_domain():
             slots=[],
             templates={},
             action_names=[],
-            form_names=["random_name", "random_name"],
+            forms=["random_name", "random_name"],
         )
 
 
@@ -644,7 +660,7 @@ def test_clean_domain_for_file():
         "actions": ["utter_default", "utter_goodbye", "utter_greet"],
         "session_config": {
             "carry_over_slots_to_new_session": True,
-            "session_expiration_time": 0,
+            "session_expiration_time": DEFAULT_SESSION_EXPIRATION_TIME_IN_MINUTES,
         },
     }
 
@@ -705,17 +721,17 @@ def test_add_knowledge_base_slots(default_domain):
     "input_domain, expected_session_expiration_time, expected_carry_over_slots",
     [
         (
-            """session_config:
-    session_expiration_time: 0
+            f"""session_config:
+    session_expiration_time: {DEFAULT_SESSION_EXPIRATION_TIME_IN_MINUTES}
     carry_over_slots_to_new_session: true""",
-            0,
+            DEFAULT_SESSION_EXPIRATION_TIME_IN_MINUTES,
             True,
         ),
-        ("", 0, True),
+        ("", DEFAULT_SESSION_EXPIRATION_TIME_IN_MINUTES, True),
         (
             """session_config:
     carry_over_slots_to_new_session: false""",
-            0,
+            DEFAULT_SESSION_EXPIRATION_TIME_IN_MINUTES,
             False,
         ),
         (
@@ -725,7 +741,7 @@ def test_add_knowledge_base_slots(default_domain):
             20.2,
             False,
         ),
-        ("""session_config: {}""", 0, True),
+        ("""session_config: {}""", DEFAULT_SESSION_EXPIRATION_TIME_IN_MINUTES, True),
     ],
 )
 def test_session_config(
@@ -765,7 +781,7 @@ def test_are_sessions_enabled(session_config: SessionConfig, enabled: bool):
 
 
 def test_domain_utterance_actions_deprecated_templates():
-    new_yaml = """actions:
+    new_yaml = f"""actions:
 - utter_greet
 - utter_goodbye
 config:
@@ -780,10 +796,10 @@ templates:
   - text: bye!
 session_config:
   carry_over_slots_to_new_session: true
-  session_expiration_time: 60
-slots: {}"""
+  session_expiration_time: {DEFAULT_SESSION_EXPIRATION_TIME_IN_MINUTES}
+slots: {{}}"""
 
-    old_yaml = """config:
+    old_yaml = f"""config:
   store_entities_as_slots: true
 entities: []
 forms: []
@@ -795,8 +811,8 @@ responses:
   - text: bye!
 session_config:
   carry_over_slots_to_new_session: true
-  session_expiration_time: 60
-slots: {}"""
+  session_expiration_time: {DEFAULT_SESSION_EXPIRATION_TIME_IN_MINUTES}
+slots: {{}}"""
 
     old_domain = Domain.from_yaml(old_yaml)
     new_domain = Domain.from_yaml(new_yaml)