+++ tests/core/training/story_reader/test_yaml_story_reader.py
@@ -1,8 +1,12 @@
+from typing import Text
+
 import pytest
 
+from rasa.constants import LATEST_TRAINING_DATA_FORMAT_VERSION
 from rasa.core import training
 from rasa.core.domain import Domain
-from rasa.core.events import ActionExecuted, UserUttered, SlotSet
+from rasa.core.training import loading
+from rasa.core.events import ActionExecuted, UserUttered, SlotSet, Form
 from rasa.core.interpreter import RegexInterpreter
 from rasa.core.training.story_reader.yaml_story_reader import YAMLStoryReader
 from rasa.utils import io as io_utils
@@ -86,18 +90,22 @@ async def test_can_read_test_story_with_entities_without_value(default_domain: D
     assert trackers[0].events[-1] == ActionExecuted("action_listen")
 
 
-async def test_is_yaml_file():
-    valid_yaml_file = "data/test_yaml_stories/stories.yml"
-    valid_markdown_file = "data/test_stories/stories.md"
-
-    assert YAMLStoryReader.is_yaml_story_file(valid_yaml_file)
-    assert not YAMLStoryReader.is_yaml_story_file(valid_markdown_file)
+@pytest.mark.parametrize(
+    "file,is_yaml_file",
+    [
+        ("data/test_yaml_stories/stories.yml", True),
+        ("data/test_stories/stories.md", False),
+        ("data/test_yaml_stories/rules_without_stories.yml", True),
+    ],
+)
+async def test_is_yaml_file(file: Text, is_yaml_file: bool):
+    assert YAMLStoryReader.is_yaml_story_file(file) == is_yaml_file
 
 
 async def test_yaml_intent_with_leading_slash_warning(default_domain: Domain):
     yaml_file = "data/test_wrong_yaml_stories/intent_with_leading_slash.yml"
 
-    with pytest.warns(UserWarning):
+    with pytest.warns(UserWarning) as record:
         tracker = await training.load_data(
             yaml_file,
             default_domain,
@@ -105,20 +113,83 @@ async def test_yaml_intent_with_leading_slash_warning(default_domain: Domain):
             tracker_limit=1000,
             remove_duplicates=False,
         )
-        assert tracker[0].latest_message == UserUttered("simple", {"name": "simple"})
+
+    # one for leading slash, one for missing version
+    assert len(record) == 2
+
+    assert tracker[0].latest_message == UserUttered("simple", {"name": "simple"})
 
 
 async def test_yaml_wrong_yaml_format_warning(default_domain: Domain):
     yaml_file = "data/test_wrong_yaml_stories/wrong_yaml.yml"
 
     with pytest.warns(UserWarning):
-        await training.load_data(
+        _ = await training.load_data(
             yaml_file,
             default_domain,
             use_story_concatenation=False,
             tracker_limit=1000,
             remove_duplicates=False,
         )
+
+
+async def test_read_rules_with_stories(default_domain: Domain):
+
+    yaml_file = "data/test_yaml_stories/stories_and_rules.yml"
+
+    steps = await loading.load_data_from_files(
+        [yaml_file], default_domain, RegexInterpreter()
+    )
+
+    ml_steps = [s for s in steps if not s.is_rule]
+    rule_steps = [s for s in steps if s.is_rule]
+
+    # this file contains three rules and three ML stories
+    assert len(ml_steps) == 3
+    assert len(rule_steps) == 3
+
+    assert rule_steps[0].block_name == "rule 1"
+    assert rule_steps[1].block_name == "rule 2"
+    assert rule_steps[2].block_name == "rule 3"
+
+    assert ml_steps[0].block_name == "simple_story_without_checkpoint"
+    assert ml_steps[1].block_name == "simple_story_with_only_start"
+    assert ml_steps[2].block_name == "simple_story_with_only_end"
+
+
+async def test_read_rules_without_stories(default_domain: Domain):
+
+    yaml_file = "data/test_yaml_stories/rules_without_stories.yml"
+
+    steps = await loading.load_data_from_files(
+        [yaml_file], default_domain, RegexInterpreter()
+    )
+
+    ml_steps = [s for s in steps if not s.is_rule]
+    rule_steps = [s for s in steps if s.is_rule]
+
+    # this file contains three rules and no ML stories
+    assert len(ml_steps) == 0
+    assert len(rule_steps) == 3
+
+    assert rule_steps[0].block_name == "rule 1"
+    assert rule_steps[1].block_name == "rule 2"
+    assert rule_steps[2].block_name == "rule 3"
+
+    # inspect the first rule and make sure all events were picked up correctly
+    events = rule_steps[0].events
+
+    assert len(events) == 5
+
+    assert events[0] == Form("loop_q_form")
+    assert events[1] == SlotSet("requested_slot", "some_slot")
+    assert events[2] == ActionExecuted("...")
+    assert events[3] == UserUttered(
+        "inform",
+        {"name": "inform", "confidence": 1.0},
+        [{"entity": "some_slot", "value": "bla"}],
+    )
+    assert events[4] == ActionExecuted("loop_q_form")
 
 
 async def test_warning_if_intent_not_in_domain(default_domain: Domain):
@@ -133,17 +204,21 @@ async def test_warning_if_intent_not_in_domain(default_domain: Domain):
     reader = YAMLStoryReader(RegexInterpreter(), default_domain)
     yaml_content = io_utils.read_yaml(stories)
 
-    with pytest.warns(UserWarning):
+    with pytest.warns(UserWarning) as record:
         reader.read_from_parsed_yaml(yaml_content)
+
+    # one for missing intent, one for missing version
+    assert len(record) == 2
 
 
 async def test_no_warning_if_intent_in_domain(default_domain: Domain):
-    stories = """
-    stories:
-    - story: I am fine 💥
-      steps:
-      - intent: greet
-    """
+    stories = (
+        f'version: "{LATEST_TRAINING_DATA_FORMAT_VERSION}"\n'
+        f"stories:\n"
+        f"- story: I am fine 💥\n"
+        f"  steps:\n"
+        f"  - intent: greet"
+    )
 
     reader = YAMLStoryReader(RegexInterpreter(), default_domain)
     yaml_content = io_utils.read_yaml(stories)
@@ -151,4 +226,4 @@ async def test_no_warning_if_intent_in_domain(default_domain: Domain):
     with pytest.warns(None) as record:
         reader.read_from_parsed_yaml(yaml_content)
 
-    assert len(record) == 0
+    assert not len(record)