+++ tests/nlu/tokenizers/test_whitespace_tokenizer.py
@@ -1,5 +1,7 @@
 import pytest
 
+from rasa.nlu.components import UnsupportedLanguageError
+from rasa.nlu.config import RasaNLUModelConfig
 from rasa.nlu.constants import TOKENS_NAMES, TEXT, INTENT
 from rasa.nlu.training_data import TrainingData, Message
 from rasa.nlu.tokenizers.whitespace_tokenizer import WhitespaceTokenizer
@@ -99,25 +101,6 @@ def test_custom_intent_symbol(text, expected_tokens):
     assert [t.text for t in message.get(TOKENS_NAMES[INTENT])] == expected_tokens
 
 
-@pytest.mark.parametrize(
-    "text, component_config, expected_tokens",
-    [
-        ("Forecast for LUNCH", {}, ["Forecast", "for", "LUNCH"]),
-        ("Forecast for LUNCH", {"case_sensitive": False}, ["forecast", "for", "lunch"]),
-        ("Forecast for LUNCH", {"case_sensitive": True}, ["Forecast", "for", "LUNCH"]),
-    ],
-)
-def test_whitespace_with_case(text, component_config, expected_tokens):
-
-    tk = WhitespaceTokenizer(component_config)
-
-    message = Message(text)
-
-    tokens = tk.tokenize(message, attribute=TEXT)
-
-    assert [t.text for t in tokens] == expected_tokens
-
-
 def test_whitespace_training(supervised_embeddings_config):
     examples = [
         Message(
@@ -140,16 +123,39 @@ def test_whitespace_training(supervised_embeddings_config):
         ),
     ]
 
-    component_config = {"case_sensitive": False}
-    tk = WhitespaceTokenizer(component_config)
+    tk = WhitespaceTokenizer()
 
     tk.train(TrainingData(training_examples=examples), supervised_embeddings_config)
 
-    assert examples[0].data.get(TOKENS_NAMES[TEXT])[0].text == "any"
-    assert examples[0].data.get(TOKENS_NAMES[TEXT])[1].text == "mexican"
+    assert examples[0].data.get(TOKENS_NAMES[TEXT])[0].text == "Any"
+    assert examples[0].data.get(TOKENS_NAMES[TEXT])[1].text == "Mexican"
     assert examples[0].data.get(TOKENS_NAMES[TEXT])[2].text == "restaurant"
     assert examples[0].data.get(TOKENS_NAMES[TEXT])[3].text == "will"
     assert examples[0].data.get(TOKENS_NAMES[TEXT])[4].text == "do"
-    assert examples[1].data.get(TOKENS_NAMES[TEXT])[0].text == "i"
+    assert examples[1].data.get(TOKENS_NAMES[TEXT])[0].text == "I"
     assert examples[1].data.get(TOKENS_NAMES[TEXT])[1].text == "want"
-    assert examples[1].data.get(TOKENS_NAMES[TEXT])[2].text == "tacos"
+    assert examples[1].data.get(TOKENS_NAMES[TEXT])[2].text == "Tacos"
+
+
+def test_whitespace_does_not_throw_error():
+    import rasa.utils.io as io_utils
+
+    texts = io_utils.read_json_file("data/test_tokenizers/naughty_strings.json")
+
+    tk = WhitespaceTokenizer()
+
+    for text in texts:
+        tk.tokenize(Message(text), attribute=TEXT)
+
+
+@pytest.mark.parametrize("language, error", [("en", False), ("zh", True)])
+def test_whitespace_language_suuport(language, error, component_builder):
+    config = RasaNLUModelConfig(
+        {"language": language, "pipeline": [{"name": "WhitespaceTokenizer"}]}
+    )
+
+    if error:
+        with pytest.raises(UnsupportedLanguageError):
+            component_builder.create_component({"name": "WhitespaceTokenizer"}, config)
+    else:
+        component_builder.create_component({"name": "WhitespaceTokenizer"}, config)