+++ CHANGELOG.rst
@@ -17,6 +17,38 @@ This project adheres to `Semantic Versioning`_ starting with version 1.0.
 
 .. towncrier release notes start
 
+[1.10.8] - 2020-07-15
+^^^^^^^^^^^^^^^^^^^^^
+
+Bugfixes
+--------
+- `#6075 <https://github.com/rasahq/rasa/issues/6075>`_: Add 'Access-Control-Expose-Headers' for 'filename' header
+- `#6137 <https://github.com/rasahq/rasa/issues/6137>`_: Fixed a bug where an invalid language variable prevents rasa from finding training examples when importing Dialogflow data.
+
+
+[1.10.7] - 2020-07-07
+^^^^^^^^^^^^^^^^^^^^^
+
+Features
+--------
+- `#6150 <https://github.com/rasahq/rasa/issues/6150>`_: Add ``not_supported_language_list`` to component to be able to define languages that a component can NOT handle.
+
+  ``WhitespaceTokenizer`` is not able to process languages which are not separated by whitespace. ``WhitespaceTokenizer``
+  will throw an error if it is used with Chinese, Japanese, and Thai.
+
+Bugfixes
+--------
+- `#6150 <https://github.com/rasahq/rasa/issues/6150>`_: ``WhitespaceTokenizer`` only removes emoji if complete token matches emoji regex.
+
+
+[1.10.6] - 2020-07-06
+^^^^^^^^^^^^^^^^^^^^^
+
+Bugfixes
+--------
+- `#6143 <https://github.com/rasahq/rasa/issues/6143>`_: Prevent ``WhitespaceTokenizer`` from outputting empty list of tokens.
+
+
 [1.10.5] - 2020-07-02
 ^^^^^^^^^^^^^^^^^^^^^
 
@@ -62,7 +94,7 @@ Bugfixes
 Bugfixes
 --------
 - `#5521 <https://github.com/rasahq/rasa/issues/5521>`_: Responses used in ResponseSelector now support new lines with explicitly adding ``\n`` between them.
-- `#5758 <https://github.com/rasahq/rasa/issues/5758>`_: Fixed a bug in `rasa export <https://rasa.com/docs/rasa-x/installation-and-setup/existing-deployment/#migrate-conversations>`_ (:ref:`section_export`) which caused Rasa Open Source to only migrate conversation events from the last :ref:`session_config`.
+- `#5758 <https://github.com/rasahq/rasa/issues/5758>`_: Fixed a bug in `rasa export <https://rasa.com/docs/rasa-x/installation-and-setup/deploy#connect-rasa-deployment>`_ (:ref:`section_export`) which caused Rasa Open Source to only migrate conversation events from the last :ref:`session_config`.
 
 
 [1.10.1] - 2020-05-15